Bosstag changes for peeps (#6275)
* aura effect line of sight
* allow avenger boss to be damaged
* swing boss angle
* fix surge data

+++ plugins/paper/src/main/java/com/playmonumenta/plugins/bosses/bosses/NovaBoss.java
@@ -101,7 +101,7 @@ public final class NovaBoss extends BossAbilityGroup {
 			p.NEED_LINE_OF_SIGHT = false;
 		}
 
-		Spell spell = new SpellBaseAoE(plugin, boss, (int) p.TARGETS.getRange(), p.DURATION, p.COOLDOWN, p.CAN_MOVE, p.NEED_LINE_OF_SIGHT, p.SOUND_CHARGE, p.SOUND_CHARGE_VOLUME, p.SOUND_CHARGE_FREQUENCY) {
+		Spell spell = new SpellBaseAoE(plugin, boss, p.TARGETS.getRange(), p.DURATION, p.COOLDOWN, p.CAN_MOVE, p.NEED_LINE_OF_SIGHT, p.SOUND_CHARGE, p.SOUND_CHARGE_VOLUME, p.SOUND_CHARGE_FREQUENCY) {
 			@Override
 			protected void chargeAuraAction(Location loc) {
 				p.PARTICLE_AIR.spawn(boss, loc, p.TARGETS.getRange() / 2, p.TARGETS.getRange() / 2, p.TARGETS.getRange() / 2, 0.05);
